Update import statements in all files under module with respect to the refactoring
