+++ models/cart.py
@@ -1,5 +1,5 @@
-from order import Order
-from payment import Payment
+from models.order import Order
+from models.payment import Payment
 
 
 class Cart:
@@ -16,5 +16,8 @@ class Cart:
         total_price = sum(item.price for item in self.items)
         order = Order(customer, self.items, total_price)
         payment = Payment()
-        payment.process_a_payment(order, payment_info)
+        payment.process_payment(order, payment_info)
         return order
+
+    def total_price(self, customer, payment_info):
+        return sum(item.price for item in self.items)
